infra: change folder structure
