+++ lib/feature/home/domain/entities/task.dart
@@ -1,0 +1,6 @@
+class Task {
+  final String title;
+  bool isCompleted;
+
+  Task({required this.title, this.isCompleted = false});
+}
